# specs/001-phase-2-saas/contracts/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Todo App API",
    "description": "Premium Web SaaS API for Phase 2 Todo Application with Authentication, Todo Management, User Profiles, and AI Features",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    },
    {
      "url": "https://todo-app-api.onrender.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/todos": {
      "get": {
        "summary": "List todos",
        "description": "Get user's todos with filtering, searching, and sorting",
        "operationId": "list_todos",
        "tags": ["Todos"],
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "$ref": "#/components/parameters/SkipParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          },
          {
            "$ref": "#/components/parameters/StatusFilter"
          },
          {
            "$ref": "#/components/parameters/PriorityFilter"
          },
          {
            "$ref": "#/components/parameters/SearchParam"
          },
          {
            "$ref": "#/components/parameters/SortByParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Todos retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoListResponse"
                }
              }
            },
            "401": {
              "$ref": "#/components/responses/Unauthorized"
            }
          }
        }
      },
      "post": {
        "summary": "Create todo",
        "description": "Create a new todo for authenticated user",
        "operationId": "create_todo",
        "tags": ["Todos"],
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Todo created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoResponse"
                }
              }
            },
            "400": {
              "$ref": "#/components/responses/BadRequest"
            },
            "401": {
              "$ref": "#/components/responses/Unauthorized"
            },
            "422": {
              "$ref": "#/components/responses/ValidationError"
            }
          }
        }
      }
    },
    "/todos/{id}": {
      "get": {
        "summary": "Get todo by ID",
        "description": "Retrieve a specific todo",
        "operationId": "get_todo",
        "tags": ["Todos"],
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "$ref": "#/components/parameters/TodoIdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Todo retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoResponse"
                }
              }
            },
            "401": {
              "$ref": "#/components/responses/Unauthorized"
            },
            "403": {
              "$ref": "#/components/responses/Forbidden"
            },
            "404": {
              "$ref": "#/components/responses/NotFound"
            }
          }
        }
      },
      "put": {
        "summary": "Update todo",
        "description": "Update a todo's details",
        "operationId": "update_todo",
        "tags": ["Todos"],
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "$ref": "#/components/parameters/TodoIdParam"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Todo updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoResponse"
                }
              }
            },
            "400": {
              "$ref": "#/components/responses/BadRequest"
            },
            "401": {
              "$ref": "#/components/responses/Unauthorized"
            },
            "403": {
              "$ref": "#/components/responses/Forbidden"
            },
            "404": {
              "$ref": "#/components/responses/NotFound"
            },
            "422": {
              "$ref": "#/components/responses/ValidationError"
            }
          }
        }
      },
      "delete": {
        "summary": "Delete todo",
        "description": "Delete a todo",
        "operationId": "delete_todo",
        "tags": ["Todos"],
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "$ref": "#/components/parameters/TodoIdParam"
          }
        ],
        "responses": {
          "204": {
            "description": "Todo deleted"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
            "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/todos/{id}/complete": {
      "patch": {
        "summary": "Toggle todo completion",
        "description": "Mark todo as completed or pending",
        "operationId": "toggle_todo_complete",
        "tags": ["Todos"],
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "$ref": "#/components/parameters/TodoIdParam"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["completed"],
                "properties": {
                  "completed": {
                    "type": "boolean",
                    "description": "True to mark as completed, false to mark as pending"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Todo status updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoResponse"
                }
              }
            },
            "401": {
              "$ref": "#/components/responses/Unauthorized"
            },
            "403": {
              "$ref": "#/components/responses/Forbidden"
            },
            "404": {
              "$ref": "#/components/responses/NotFound"
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token from httpOnly cookie or Authorization header"
      }
    },
    "schemas": {
...
      "TodoCreateRequest": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500,
            "description": "Todo title",
            "example": "Complete project documentation"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "Todo description",
            "example": "Write comprehensive docs for Phase 2"
          },
          "priority": {
            "type": "string",
            "enum": ["low", "medium", "high"],
            "default": "medium",
            "example": "high"
          },
          "due_date": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "example": "2025-01-30"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true,
            "example": ["work", "urgent"]
          }
        }
      },
      "TodoUpdateRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "priority": {
            "type": "string",
            "enum": ["low", "medium", "high"]
          },
          "due_date": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        }
      },
      "TodoResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": ["pending", "completed"]
          },
          "priority": {
            "type": "string",
            "enum": ["low", "medium", "high"]
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "due_date": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "TodoListResponse": {
        "type": "object",
        "properties": {
          "todos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TodoResponse"
            }
          },
          "total": {
            "type": "integer",
            "description": "Total count of user's todos",
            "example": 150
          },
          "skip": {
            "type": "integer",
            "example": 0
          },
          "limit": {
            "type": "integer",
            "example": 20
          },
          "has_more": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "string",
            "example": "Invalid credentials"
          }
        }
      }
    },
    "parameters": {
      "TodoIdParam": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "description": "Todo ID",
        "example": "550e8400-e29b-41d4-a716-446655440000"
      },
      "SkipParam": {
        "name": "skip",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 0,
          "minimum": 0
        },
        "description": "Number of todos to skip (pagination)",
        "example": 0
      },
      "LimitParam": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 20,
          "minimum": 1,
          "maximum": 100
        },
        "description": "Number of todos to return (pagination)",
        "example": 20
      },
      "StatusFilter": {
        "name": "status",
        "in": "query",
        "schema": {
          "type": "string",
          "enum": ["all", "pending", "completed"],
          "default": "all"
        },
        "description": "Filter by todo status",
        "example": "pending"
      },
      "PriorityFilter": {
        "name": "priority",
        "in": "query",
        "schema": {
          "type": "string",
          "enum": ["low", "medium", "high"]
        },
        "description": "Filter by todo priority",
        "example": "high"
      },
      "SearchParam": {
        "name": "search",
        "in": "query",
        "schema": {
          "type": "string",
          "minLength": 2
        },
        "description": "Search in todo title and description",
        "example": "documentation"
      },
      "SortByParam": {
        "name": "sort_by",
        "in": "query",
        "schema": {
          "type": "string",
          "enum": ["created_at", "due_date", "priority"],
          "default": "created_at"
        },
        "description": "Sort todos by field",
        "example": "due_date"
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "detail": "Invalid input data"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "detail": "Invalid authentication credentials"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "detail": "You don't have permission to access this resource"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "detail": "Todo not found"
            }
          }
        }
      },
...
      "ValidationError": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "detail": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "loc": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "msg": {
                        "type": "string"
                      },
                      "type": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    }
  },
...
}